Changed the type of input and output in avgpool kernel

+++ OpenCL_Kernels/host_code/CustomKernels.cl
@@ -181,14 +181,14 @@ __kernel void MaxPool(__global int * restrict img,
  * Used for reducing the dimension of images.
  */
 
-__kernel void AvgPool(__global int * restrict input, 
+__kernel void AvgPool(__global double * restrict input, 
 			int number_of_image_rows, 
 			int  number_of_image_cols,
 			int number_of_filters,
 			int kernel_size, 
 			int stride,
 			int number_of_images,
-			__global int * restrict output){
+			__global double * restrict output){
 	double avgpool[200]; 
 	int image_size = number_of_image_rows * number_of_image_cols * number_of_filters * number_of_images;
 	int avg=0, i, oindex=0,count=kernel_size, s=kernel_size, k, f, startIndex=0, endIndex=s, imageIndex=1, j=0;
